Revert hook changes
