+++ application/ui/src/features/models/hooks/api/use-get-dataset-revisions.ts
@@ -1,0 +1,13 @@
+// Copyright (C) 2025 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+
+export const useGetDatasetRevisions = () => {
+    return {
+        data: [
+            { id: '1', name: 'Dataset 1' },
+            { id: '2', name: 'Dataset 2' },
+            { id: '3', name: 'Dataset 3' },
+            { id: '4', name: 'Dataset 4' },
+        ],
+    };
+};
