+++ application/ui/src/features/models/hooks/api/use-get-model-architectures.hook.ts
@@ -1,0 +1,18 @@
+// Copyright (C) 2025 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+
+import { useProject } from 'hooks/api/project.hook';
+
+import { $api } from '../../../../api/client';
+
+export const useGetTaskModelArchitectures = () => {
+    const { data: projectData } = useProject();
+
+    return $api.useSuspenseQuery('get', '/api/model_architectures', {
+        params: {
+            query: {
+                task: projectData.task.task_type,
+            },
+        },
+    });
+};
