+++ application/ui/src/features/models/hooks/api/use-get-model.hook.ts
@@ -1,0 +1,17 @@
+// Copyright (C) 2025 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+
+import { useProjectIdentifier } from 'hooks/use-project-identifier.hook';
+
+import { $api } from '../../../../api/client';
+
+export const useGetModel = (modelId: string | null | undefined) => {
+    const projectId = useProjectIdentifier();
+
+    return $api.useQuery(
+        'get',
+        '/api/projects/{project_id}/models/{model_id}',
+        { params: { path: { project_id: projectId, model_id: String(modelId) } } },
+        { enabled: Boolean(modelId) }
+    );
+};
