+++ application/ui/src/features/models/hooks/api/use-get-training-devices.ts
@@ -1,0 +1,8 @@
+// Copyright (C) 2025 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+
+import { $api } from '../../../../api/client';
+
+export const useGetTrainingDevices = () => {
+    return $api.useSuspenseQuery('get', '/api/system/devices/training');
+};
