+++ application/ui/src/features/models/model-listing/current-model-training/training-model-row.component.tsx
@@ -7,7 +7,7 @@ import duration from 'dayjs/plugin/duration';
 
 import { ReactComponent as ThumbsUp } from '../../../../assets/icons/thumbs-up.svg';
 import type { Job } from '../../../../constants/shared-types';
-import { useGetModels } from '../../../../hooks/api/models.hook';
+import { useGetModels } from '../../hooks/api/use-get-models.hook';
 import { GRID_COLUMNS } from '../constants';
 import { BottomProgressBar } from './bottom-progress-bar.component';
