+++ application/ui/src/features/models/train-model/train-model-provider.component.tsx
@@ -11,12 +11,10 @@ import {
     RecommendedModelArchitectures,
     TrainingDevice,
 } from '../../../constants/shared-types';
-import {
-    useGetActiveModelArchitectureId,
-    useGetDatasetRevisions,
-    useGetTaskModelArchitectures,
-    useGetTrainingDevices,
-} from '../../../hooks/api/models.hook';
+import { useGetActiveModelArchitectureId } from '../hooks/api/use-get-active-model-architecture-id.hook';
+import { useGetDatasetRevisions } from '../hooks/api/use-get-dataset-revisions';
+import { useGetTaskModelArchitectures } from '../hooks/api/use-get-model-architectures.hook';
+import { useGetTrainingDevices } from '../hooks/api/use-get-training-devices';
 
 type TrainModelContextProps = {
     modelArchitectures: ModelArchitectureWithPerformanceCategory[];
